Move _V5HTTPManager to module Extract auth logic into separate class
