+++ pybit/_v5_spot_leverage_token.py
@@ -1,5 +1,5 @@
-from ._http_manager import _V5HTTPManager
-from .spot_leverage_token import SpotLeverageToken
+from pybit._http_manager import _V5HTTPManager
+from pybit.spot_leverage_token import SpotLeverageToken
 
 
 class SpotLeverageHTTP(_V5HTTPManager):
